requirements.txt
anthropic>=0.86.0
streamlit>=1.55.0
langchain>=1.2.13
langchain-anthropic>=1.4.0
langchain-core>=1.2.20
langgraph>=1.1.3
mcp>=1.27.0
pyyaml>=6.0.0
